finally fixed json file problem

+++ models/engine/file_storage.py
@@ -20,15 +20,16 @@ class FileStorage:
         self.__objects[key] = obj
 
     def save(self, obj):
-        self.new(obj)
-        srl_objs = {}
-        for key, obj in self.__objects.items():
-            srl_objs[key] = obj.to_dict()
-        print("cho")
-        print(self)
-        print("end")
-        with open(self.__file_path, 'w') as file:
-            json.dump(srl_objs, file)
+        if os.path.isfile(self.__file_path):
+            self.new(obj)
+            srl_objs = {}
+            for key, obj in self.__objects.items():
+                srl_objs[key] = obj.to_dict()
+            print("cho")
+            print(self)
+            print("end")
+            with open(self.__file_path, 'w') as file:
+                json.dump(srl_objs, file)
 
     def reload(self):
         if path.isfile(self.__file_path):
